adding suggested test cases
